+++ lib/rust/parser/src/syntax/token/operator.rs
@@ -45,11 +45,7 @@ impl OperatorProperties {
     }
 
     /// Return a copy of this operator, with the given binary infix precedence.
-    pub fn with_binary_infix_precedence(self, value: u8) -> Self {
-        let precedence = Precedence { value };
-        debug_assert!(precedence > Precedence::min());
-        debug_assert!(value & 0x80 == 0);
-        debug_assert!((value + 1) & 0x80 == 0);
+    pub fn with_binary_infix_precedence(self, precedence: Precedence) -> Self {
         Self { binary_infix_precedence: Some(precedence), ..self }
     }
 
@@ -159,7 +155,7 @@ impl<'s> TokenOperatorProperties for Token<'s> {
 impl HasOperatorProperties for variant::AssignmentOperator {
     fn operator_properties(&self) -> OperatorProperties {
         OperatorProperties {
-            binary_infix_precedence: Some(Precedence { value: 1 }),
+            binary_infix_precedence: Some(Precedence::Assignment),
             lhs_section_termination: Some(SectionTermination::Unwrap),
             is_right_associative: true,
             is_compile_time: true,
@@ -171,9 +167,8 @@ impl HasOperatorProperties for variant::AssignmentOperator {
 impl HasOperatorProperties for variant::TypeAnnotationOperator {
     fn operator_properties(&self) -> OperatorProperties {
         OperatorProperties {
-            binary_infix_precedence: Some(Precedence { value: 2 }),
+            binary_infix_precedence: Some(Precedence::TypeAnnotation),
             lhs_section_termination: Some(SectionTermination::Reify),
-            is_right_associative: true,
             is_compile_time: true,
             rhs_is_non_expression: true,
             ..default()
@@ -184,7 +179,7 @@ impl HasOperatorProperties for variant::TypeAnnotationOperator {
 impl HasOperatorProperties for variant::ArrowOperator {
     fn operator_properties(&self) -> OperatorProperties {
         OperatorProperties {
-            binary_infix_precedence: Some(Precedence { value: 2 }),
+            binary_infix_precedence: Some(Precedence::Arrow),
             lhs_section_termination: Some(SectionTermination::Unwrap),
             is_right_associative: true,
             is_compile_time: true,
@@ -196,7 +191,7 @@ impl HasOperatorProperties for variant::ArrowOperator {
 impl HasOperatorProperties for variant::AnnotationOperator {
     fn operator_properties(&self) -> OperatorProperties {
         OperatorProperties {
-            unary_prefix_precedence: Some(Precedence::max()),
+            unary_prefix_precedence: Some(Precedence::Annotation),
             is_right_associative: true,
             is_compile_time: true,
             rhs_is_non_expression: true,
@@ -220,7 +215,7 @@ impl HasOperatorProperties for variant::NegationOperator {
     fn operator_properties(&self) -> OperatorProperties {
         OperatorProperties {
             is_value_operation: true,
-            unary_prefix_precedence: Some(Precedence::unary_minus()),
+            unary_prefix_precedence: Some(Precedence::Negation),
             ..default()
         }
     }
@@ -238,7 +233,7 @@ impl HasOperatorProperties for variant::LambdaOperator {
 
 impl HasOperatorProperties for variant::DotOperator {
     fn operator_properties(&self) -> OperatorProperties {
-        OperatorProperties { binary_infix_precedence: Some(Precedence { value: 80 }), ..default() }
+        OperatorProperties { binary_infix_precedence: Some(Precedence::Application), ..default() }
     }
 }
 
@@ -256,7 +251,7 @@ impl HasOperatorProperties for variant::SuspensionOperator {
 impl HasOperatorProperties for variant::CommaOperator {
     fn operator_properties(&self) -> OperatorProperties {
         OperatorProperties {
-            binary_infix_precedence: Some(Precedence { value: 1 }),
+            binary_infix_precedence: Some(Precedence::Assignment),
             is_compile_time: true,
             rhs_is_non_expression: true,
             ..default()
@@ -267,50 +262,56 @@ impl HasOperatorProperties for variant::CommaOperator {
 /// Value that can be compared to determine which operator will bind more tightly within an
 /// expression.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-pub struct Precedence {
-    /// A numeric value determining precedence order.
-    pub(super) value: u8,
+#[repr(u8)]
+#[allow(missing_docs)]
+pub enum Precedence {
+    /// A value that is lower than the precedence of any operator.
+    Min = 0,
+    Assignment,
+    TypeAnnotation,
+    Arrow,
+    Not,
+    Logical,
+    Equality,
+    Functional,
+    BitwiseOr,
+    BitwiseAnd,
+    Inequality,
+    Addition,
+    Multiplication,
+    Exponentiation,
+    OtherUserOperator,
+    Negation,
+    Application,
+    Annotation,
+    /// A value that is higher than the precedence of any operator.
+    Max,
 }
 
 impl Precedence {
     /// Return a precedence that is lower than the precedence of any operator.
     pub fn min() -> Self {
-        Precedence { value: 0 }
+        Precedence::Min
     }
 
-    /// Return the precedence for any operator.
+    /// Return the lowest precedence for any operator.
     pub fn min_valid() -> Self {
-        Precedence { value: 1 }
+        debug_assert_eq!(Precedence::Assignment as u8, Precedence::Min as u8 + 1);
+        Precedence::Assignment
     }
 
     /// Return a precedence that is not lower than any other precedence.
     pub fn max() -> Self {
-        Precedence { value: 100 }
-    }
-
-    /// Return the precedence of application.
-    pub fn application() -> Self {
-        Precedence { value: 80 }
-    }
-
-    /// Return the precedence of @annotations.
-    pub fn annotation() -> Self {
-        Precedence { value: 79 }
-    }
-
-    /// Return the precedence of unary minus.
-    pub fn unary_minus() -> Self {
-        Precedence { value: 79 }
-    }
-
-    /// Return the precedence of unary minus when applied to a numeric literal.
-    pub fn unary_minus_numeric_literal() -> Self {
-        Precedence { value: 80 }
+        Precedence::Max
     }
 
     /// Return the value as a number.
     pub fn into_u8(self) -> u8 {
-        self.value
+        let value = self as u8;
+        debug_assert!(value > Precedence::Min as u8);
+        debug_assert!(value & 0x80 == 0);
+        debug_assert!((value + 1) & 0x80 == 0);
+        value
     }
 }