+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - [Visualizations on components are slightly transparent when not
   focused][11582].
 - [New design for vector-editing widget][11620]
+- [The `:` type operator can now be chained][11671]
 
 [11151]: https://github.com/enso-org/enso/pull/11151
 [11271]: https://github.com/enso-org/enso/pull/11271
@@ -60,6 +61,7 @@
 [11612]: https://github.com/enso-org/enso/pull/11612
 [11582]: https://github.com/enso-org/enso/pull/11582
 [11620]: https://github.com/enso-org/enso/pull/11620
+[11671]: https://github.com/enso-org/enso/pull/11671
 
 #### Enso Standard Library